chore: move retry to utils folder
